user_auth_project/requirements.txt
Flask==2.3.3
firebase-admin==6.4.0
google-cloud-datastore==2.17.0
gunicorn==21.2.0
